+++ plugin_info/configuration.php
@@ -56,4 +56,3 @@ if (!isConnect()) {
 
     </fieldset>
 </form>
-