+++ docs/en_US/changelog.md
@@ -1,4 +1,7 @@
 #Changelog : Netatmo OpenData
 
+## 23 may 2020
+- Add new data  : Gust Strength,  Gust Angle, Rain on the last hour and rain for the day
+
 ## 16 may 2020
 - First release in 'beta'