+++ docs/fr_FR/changelog.md
@@ -1,5 +1,8 @@
 #Changelog : Netatmo OpenData
 
+## 23 mai 2020
+- Ajout de nouvelles données : rafales de vent (vitesse et direction), pluie sur 1h, pluie sur la journée
+
 ## 16 mai 2020
 - 1ère version in 'beta'